+++ docs/HISTORY.rst
@@ -4,8 +4,8 @@ Changelog
 1.0.4 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- The test suite now works with Plone 5.2, Python3 and Products.ExternalEditor 3
+  [ale-rt]
 
 1.0.3 (2015-08-21)
 ------------------